web/requirements.txt
Django==3.2.7
django-filter==2.4.0
django-tables2==2.4.0
django-bootstrap3==15.0.0
django-crispy-forms==1.12.0
mysql-connector-python==8.0.26
gunicorn==20.1.0
django-autocomplete-light==3.5.1
django-leaflet==0.28.1
gdal==3.2.3
pyproj
palettable==3.3.0
